+++ lib/src/flutter_secure_share_storage.dart
@@ -22,7 +22,7 @@ class FlutterSecureShareStorage implements SecureShareStorage {
   FlutterSecureShareStorage({FlutterSecureStorage? storage})
       : _storage = storage ??
             const FlutterSecureStorage(
-              aOptions: AndroidOptions(encryptedSharedPreferences: true),
+              aOptions: AndroidOptions(),
               iOptions: IOSOptions(
                 accessibility: KeychainAccessibility.unlocked_this_device,
               ),